+++ src/lib/writers/json-value.ts
@@ -23,7 +23,7 @@ export class JsonValueWriter extends StructuredWriter {
     return null;
   }
 
-  writeBuffer(value: Uint8Array): string {
+  writeBytes(value: Uint8Array): string {
     const result: string[] = new Array(value.length);
     const len: number = value.length;
     for (let i: number = 0; i < len; i++) {